+++ src/index.ts
@@ -3,6 +3,7 @@ export * from './components/html/index';
 export * from './components/icons/index';
 export * from './components/alerts/index';
 export * from './context/theme/index';
+export * from './context/config/index';
 export * from './theme/index';
 export * from './components/VenueMapEditor/index';
 export * from './components/ElementLibraryBuilder/index';